rentalDto-rentalMapper

+++ src/main/java/com/example/car/rental/system/dto/CarDto.java
@@ -1,8 +1,8 @@
 package com.example.car.rental.system.dto;
 
-import com.example.car.rental.system.enums.AvailabilityStatusType;
-import com.example.car.rental.system.enums.FuelType;
-import com.example.car.rental.system.enums.TransmissionType;
+import com.example.car.rental.system.enums.car.AvailabilityStatusType;
+import com.example.car.rental.system.enums.car.FuelType;
+import com.example.car.rental.system.enums.car.TransmissionType;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
@@ -11,7 +11,7 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 public class CarDto {
-    private int id ;
+    private Long id ;
     private String brand;
     private String model;
     private String color;
